bots/voice-chat-detection/requirements.txt
discord.py==2.3.2
python-dotenv==1.1.0